Compatibility corrections for Tsuri Seensei 2 GH-57
Raphael-Boichot contributed this correction

During the printing, you can change the image on screen, it sends an INIT command to the serial that enters en infinite loop because busy state is not set to zero (FIXED).

You must escape (no choice) the printing menu by sending an ABORT command. It was not correctly taken into account in the emulator due to a redundancy that leads to an infinite loop.